+++ src/devcards/util/utils.cljs
@@ -1,8 +1,6 @@
 (ns devcards.util.utils
   (:require
-   [cljs.pprint :as pprint])
-  (:import
-   [goog.string StringBuffer]))
+   [cljs.pprint :as pprint]))
 
 (defn html-env? []
   (if-let [doc js/goog.global.document]
@@ -11,10 +9,15 @@
 (defn node-env? [] (not (nil? goog/nodeGlobalRequire)))
 
 (defn pprint-str [obj]
-   (let [sb (StringBuffer.)]
-     (pprint/pprint obj (StringBufferWriter. sb))
-     (str sb)))
+  ;; this is currently only to handle the
+  ;; problem of printing JavaScript Symbols
+  (try
+    (with-out-str (pprint/pprint obj))
+    (catch js/Error e1
+      (try
+        (.toString obj)
+        (catch js/Error e2
+          (str "<<Un-printable Type>>"))))))
  
  (defn pprint-code [code]
    (pprint/with-pprint-dispatch pprint/code-dispatch (pprint-str code)))
-